Adds attack mode with brute force capability
This commit introduces a new 'attack' operating mode,
allowing the program to perform various cryptanalytic
attacks.

Currently, it integrates a brute-force attack capability
and the `statrs` crate which provides statistical
functionalities such as the chi-squared distribution
used in cryptanalysis.

It updates the `args.rs` to include an `attack` option
in the command-line arguments and modifies
`main.rs` to invoke the attack when the corresponding flag is set.